# server/src/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oya Browser API",
    "description": "REST API for controlling browsers connected to the Oya Browser server. Browsers connect via WebSocket; AI tools and clients use this REST API (or MCP) to send commands.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://oyabrowser.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3100",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/control/sessions/{id}": {
      "get": {
        "summary": "Read a durable session",
        "operationId": "getApiControlSessionsId",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlSession"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/control/sessions/{id}/control": {
      "post": {
        "summary": "Transfer human control; release leaves agent paused",
        "operationId": "postApiControlSessionsIdControl",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "acquire",
                      "release",
                      "resume"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/control/sessions/{id}/input": {
      "post": {
        "summary": "Send input using the caller\u2019s human control lease",
        "operationId": "postApiControlSessionsIdInput",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string"
                  },
                  "params": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/control/sessions/{id}/ticket": {
      "post": {
        "summary": "Issue a one-use 60-second session-bound ticket",
        "operationId": "postApiControlSessionsIdTicket",
        "tags": [
          "Durable control"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or revoked credential"
          },
          "403": {
            "description": "Insufficient role"
          },
          "409": {
            "description": "Lifecycle, idempotency, or control lease conflict"
          },
          "503": {
            "description": "Storage or session owner unavailable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key passed as Bearer token. Create one from the dashboard after signing in."
      }
    },
...
    "schemas": {
      "ControlSession": {
        "type": "object",
        "required": [
          "id",
          "project",
          "state",
          "control"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "provisioning",
              "ready",
              "disconnected",
              "cleanup_pending",
              "unknown_outcome",
              "stopped",
              "failed"
            ]
          },
          "managed": {
            "type": "boolean"
          },
          "costUsd": {
            "type": "number"
          },
          "control": {
            "type": "object",
            "properties": {
              "mode": {
                "type": "string",
                "enum": [
                  "agent",
                  "human",
                  "paused"
                ]
              },
              "expiresAt": {
                "type": "integer",
                "nullable": true
              }
            }
          }
        }
      },
...
    }
  }
}